Add profiling for property accesses and use it to sort object properties
Summary:
This diff adds a jumpstart-based profiling mechanism to count the accesses to object properties, and uses it to reorder the properties within the objects in decreasing order of access counts. This jumpstart-based profile-guided approach was chosen because it allows HHVM to run with a single layout for all objects of a given class throughout a process execution. When the JIT profile data is going to be serialized, the profile translations are instrumented with counters to count the property to accesses. This counters are then serialized along with the serialized JIT profile data. When HHVM is started in JIT-deserialization mode, the property-access counters are used to decide the order of the object properties in memory. This decision is done at class-creating time, before any object from a given class is instantiated, thus allowing the property layout for all objects of a given class to remain the same during a process execution.

Note that, due to inheritance, properties can only be reordered within each level of the class hierarchy. Also, properties imported from traits can be reordered along with other properties declared in the same class where the trait is imported. Finally, the layout of redeclared properties is determined by the first class that declares them. Therefore, properly profiling redeclared properties requires keeping track of the first class that declares a property in the hierarchy (only the last class redeclaring the property was being tracked).

Reviewed By: ricklavoie

Differential Revision: D16353615

fbshipit-source-id: 3339438b090e8d3bca092092be2c61248a743d45