Decouple logical property order from physical layout in memory
Summary:
This diff decouples the logical order for properties within objects from their physical layout in memory, allowing the order of the properties to be changed in memory while preserving the observable order of the properties (e.g. when converting an object to an array). To enable this, a map from the logical order to the physical layout is maintained for each Class.

The existing code used the terms `slot` and `index` interchangeably. This diff adopts the term `slot` to mean the logical order and the `index` to be the physical order in memory.

To increase the test coverage of this change, this diff alphabetically sorts the properties in memory for debug builds.

Reviewed By: ricklavoie

Differential Revision: D15541765

fbshipit-source-id: 71c0b2ad75d4755b303bb65a498dd3eb1026b805