Fix broken xmlreader test
Summary:
This test has been broken in master for more than a day.  I bisected the problem
to D16788707, which upgraded the libxml library.  Debugging the test, I noticed
that the behavior of function xmlParserInputBufferCreateMem when passed an empty
string changed in the new library, which resulted in hhvm dropping a warning.
I'm dropping the empty string case from the test, instead of just dropping the
warning from the output, to avoid breaking the test for open-source users.

Reviewed By: ricklavoie

Differential Revision: D16824356

fbshipit-source-id: 9000e37940a7ba4cdd7ded725a12332c3553713d