Rename variables in class_get for clarity
Summary:
This function seems to have been written with static methods in mind, and then static properties were bolted on later. As a result, none of the variables make sense in context and it's hard to follow where the coercion point is (assignment to a static property). The only time we know* we are considering a static method is during the Tfun match. This diff renames the variables to clarify the member access.

*This isn't exactly true as it could be a static property with a lambda type hint, but those are not possible to write without a FIXME because lambda static initializers are a parse error.

Reviewed By: andrewjkennedy

Differential Revision: D16825294

fbshipit-source-id: 40c426d68a5da5a1dc925828d8403ff8d88eefbc