Annotations now able to be associated with corresponding disasm
Summary:
If annotations are present, then they will be used to associate additional information to the disasm.

Until we find a better overall schema, this extra information is presented as follows:
- A field "ir_annotation" is added onto each range in a region. This contains start and end addresses of the corresponding IR range, as well as an IR area and a IR instruction and block id.
- A field "ir_annotation" is also added to the top level of each translation. This is a full json-formatted dump of the IR Unit. The instruction and block ids from above can be used to index into this unit to associate additional information.

We have chosen to use this method to avoid information duplication (especially at the IR block level), while still keeping the schema from a normal tcdump without annotations.

Reviewed By: jjgriego

Differential Revision: D16609872

fbshipit-source-id: e9cc496fc190c615ba4e1e41ff5d6087082e1330