Allow waiting for requests to finish asynchronously in tests
Summary:
Presently, in LSP tests, we can choose to either wait or not wait for a response to a request before proceeding. However, if we choose not to wait for a response, we can't wait at a later time. This commit adds the functionality to tag your requests with IDs, which can be used to wait on requests later. (Note that these IDs are strings and not every request is required to have such an ID, which makes it easier to work with them than LSP request IDs).

I also removed some `wait=True` annotations, since waiting for the request to complete is actually the default.

Reviewed By: jewelpit

Differential Revision: D16799712

fbshipit-source-id: 82401964acacb17032fc1ea2009a1d61ace99d6f