Implement basic simple pessimization
Summary: This is an alternate, simpler approach to pessimization. If a type is not enforceable, at coercion points, it is wrapped in like. This diff implements the most basic bit of this feature: within a function body, the parameter and return types are pessimized. No generic constraints are modified. Subsequent diffs will pessimize function/method calls, assignment, etc, but this is enough to start testing the effects of pessimization.

Reviewed By: kmeht

Differential Revision: D16790272

fbshipit-source-id: ee03ea21802d050c2aa94b30a025e064adb53a9a