Introduce RecordArray
Summary:
This diff introduces `RecordArray`, a new `ArrayKind` which is backed by a `RecordData`-like storage. It also has a hash map  to contain arbitrary field-value pairs that are not declared in the corresponding `RecordDesc`.

This diff also introduces `NewRecordArray` bytecode to create such arrays and the `Release` method, but stubs out all other `ArrayData` methods.

Reviewed By: paulbiss

Differential Revision: D16381732

fbshipit-source-id: ed0c6e64a3becd1a61600bfa93d4413c0b737cb1