Box instances of FilePosLarge in Pos
Summary:
If we wrap RelativePaths in an Rc, then the memory required by PosImpl::Small would be just 24 bytes--8 bytes for the path pointer, and 8 bytes each for `start` and `end`.

However, because the size of an enum is that of its largest variant, we'd use more than that in practice. The size of PosImpl::Large would be 8 bytes for the pointer plus 24 bytes each for `start` and `end` (since they're three 8-byte integers each). It seems worth adding some indirection in the little-used FilePosLarge case to bring the size of a position down from 56 bytes to 24 (again, once we wrap the RelativePath in an Rc).

Reviewed By: shiqicao

Differential Revision: D16844242

fbshipit-source-id: 1b883894fe939ea5bdccaaa73b30e3ad04ef463e