Use usize rather than u64 in Pos
Summary:
Since we use `usize` elsewhere as a substitute for the OCaml `int` type (correctly, since OCaml `int`s are pointer-width), it would cause some friction for Pos APIs to deal in `u64`s instead.

It would arguably be better to use `u64`s, since it seems that File_pos_small (the OCaml module) is not written with portability for 32-bit architectures in mind (I think it would silently behave incorrectly). If we do not support 32-bit builds, then perhaps we could standardize on `u64` rather than `usize`, but we might always encounter (mild) friction (i.e., the need for casts) when it comes to third-party crates (like `ocaml`, which defines its value type using `usize`).

Considering this, I'm proposing that we use `usize` in Pos, even though it's clearly incorrect for FilePosSmall on 32-bit machines.

Reviewed By: shiqicao

Differential Revision: D16844241

fbshipit-source-id: 4bf68cdcbd5fbc1ffaad2ce908f6e1ea765176e1