Fix FFP test cases that weren't testing what they thought.
Summary:
We had 2 tests that were trying to verify that we error for keywords when they're used in places where class or function names are expected. Unfortunately, these tests didn't take into account our error recovery algorithm, and certain parsing rules that trigger at higher priority than parsing a statement that looks like a function call. The presence of these in the test caused the test to think we'd entered a classish declaration body, and we were not actually exercising the right code paths for the rest of the test - instead we were hitting errors specific to classish declaration bodies like `A class member, method, type, trait usage, trait require, xhp attribute, xhp use, or xhp category is expected here.`

Fix these tests by commenting out the problematic bits of syntax that derail the parsing of the rest of the test, and leave comments explaining why they're commented out.

Reviewed By: dabek

Differential Revision: D16782121

fbshipit-source-id: 86eeff07d21f7f1a53d3645ef3a6115bbb64235b