Remove support for the 'var' phpism.
Summary:
We have not supported the 'var' style of property declaration in the typechecker for some time, but we still allowed it in the runtime.

Remove parser rupport for it. Remove the typechecker error for it - we can treat it the same as if you put any other kind of inappropriate token in a classish body. Fix tests.

The Var token remains because it has other uses besides the phpism.

Reviewed By: kmeht

Differential Revision: D16782136

fbshipit-source-id: 3bc03f3fe1f292386c1e372c3e7b87c81e7ca51c