Enable code coverage globally or through request parameters
Summary:
With the new JIT unwinder, we'll no longer be able to throw VMSwitchMode exception, so we need to revamp how code coverage is done.

If `RuntimeOption::EvalEnableCodeCoverage == 0`, then code coverage will be ignored.
If `RuntimeOption::EvalEnableCodeCoverage == 1`, then if `enable_code_coverage=true` is passed in request params, then we force interp and turn on code coverage.
If `RuntimeOption::EvalEnableCodeCoverage > 1`, then we force interp and code coverage is on.

Reviewed By: paulbiss

Differential Revision: D16767392

fbshipit-source-id: cc9d5a4e9c81b40147ef7bdfb35be7d98066f393