Implement signature help in serverless IDE
Summary:
This stack of code is intended to fix the remaining issues with signature help and add tests to ensure that signature help continues to work as expected.

This diff enables signature help to work in serverless IDE.

However, by itself, this diff doesn't actually make the feature work.  There are other bugs with signature help that are being fixed in subsequent diffs in this stack.

Currently known bugs:
* Class methods and static methods can't be found by signature help
* Aliased namespace functions (e.g. Str\Join -> HH\Lib\Str\Join) can't be found by signature help
* Class constructors can't be found by signature help

Reviewed By: arxanas

Differential Revision: D16866666

fbshipit-source-id: bc95c83493d1c770df527490b97ab18970cf29f2