Disable typecheck_like_types
Summary: We're moving pretty strongly in the direction of simple pessimization and away from deep pessimization, so it doesn't make sense to keep around the CI test for it. I've marked them as disabled for now in case we decide we need them, but if not we can rewrite them for simple pessimization when that's closer to launch.

Reviewed By: kmeht

Differential Revision: D16875156

fbshipit-source-id: e8c93ab818694a7f8d4d62d865493f74ba6c36f0