Add Arena class to avoid calling malloc() every time we allocate
Summary:
We want to avoid calling `vec![0; size].into_boxed_slice()` every time we allocate (in the future), so this implements a simple arena that bump-allocates instead.

A chunk is just a linked-list, and `Arena` has the pointer to the last chunk in a reversed linked list.

I'm not certain what "good" metrics are here, and this was very loosely based on [arena](https://fburl.com/moypj9sc), source is [here](https://fburl.com/1quupl24).

I'm also not sure if there was a better way to accomplish this: I wanted to avoid calling anything `unsafe`, and tried to avoid `RefCell` and `Rc`, but ultimately couldn't find a way.

Reviewed By: jewelpit

Differential Revision: D16748038

fbshipit-source-id: c90646c871b6f2c4f0dbecdacd3ae8c39fc891ec