Take &mut Arena in allocation methods
Summary:
In the previous diff, we introduced BlockBuilder, a mutable view of an allocated block. In existing conversion code, we concurrently allocate and populate multiple blocks, preventing us from properly expressing that BlockBuilders mutably borrow Arenas (and, consequently, requiring a line of unsafe code in Arena).

This might be okay--concurrent BlockBuilders for a single Arena expose non-overlapping regions of memory in the Arena, and it isn't possible to get an alias to a region of memory controlled by a BlockBuilder (without finalizing it, which eliminates the mutable reference), so we might not have problems with this implementation (since mutable segments of Arena-managed memory cannot be aliased). I'm not entirely confident this is safe, though, so I would prefer to properly express that BlockBuilders mutably borrow Arenas, allowing us to get rid of this line of unsafe code in Arena.

This requires us to convert all fields of a block before allocating the block. Doing this in ocamlrep_derive required some extensive changes, so this diff also refactors ocamlrep_derive a bit.

Reviewed By: jewelpit

Differential Revision: D16817846

fbshipit-source-id: 34f5e7dc9ef9f3c5795a0b357e2f869fd7a542b3