Allocate blocks in Arena; tie Values to Arena lifetime
Summary:
To use Arena for ocamlrep values (with minimal use of `unsafe`), we need Arena to allocate chunks of `Value`s rather than `usize`s, and provide an API dealing in blocks.

To do away with the unsafety of the method `Value::as_block` (which currently breaks aliasing rules by providing a means of getting an alias to a mutable block), this diff introduces `BlockBuilder` as the mutable view of a block (making `Block` a read-only view). The new Arena APIs deal in `BlockBuilder`s--one requests that the arena allocate a block, receives a mutable view of that block, sets the fields, then converts it to a `Value` in order to use it elsewhere (e.g., as a field in another block).

Now that a `Value` refers to data owned by an `Arena` (when it is not an immediate integer) rather than owning its own data, we need to tie `Value`s to the lifetime of the `Arena` in order to prevent them from being used after the `Arena` is freed. Since `Value` is just a `usize`, we associate a lifetime using `PhantomData`.

Since blocks are now allocated only through an `Arena`, we must also introduce a new trait, `IntoOcamlRep` (rather than using `Into<ocamlrep::Value>`), which specifies a method `into_ocamlrep` that accepts the arena.

For migration purposes, the `Arena` APIs are temporarily changed to take `&self` rather than `&mut self` (using a line of unsafe code). This is fixed in the next diff (which makes some more invasive changes to uses of this API in order to allow tying `BlockBuilder` to the lifetime of this mutable borrow).

Reviewed By: jewelpit

Differential Revision: D16817834

fbshipit-source-id: ee12b844637eb95afeb260087d63df4c0f43ca65