SourceText -> Rc<SourceText>
Summary:
We've been carefully avoiding storing source text anywhere within the tree, instead opting to make the tree methods needing it accept it from outside.

This starts to get really unwieldy, so in the first few diffs of this stack I change it. To avoid storing the copy of same source text in possibly multiple nodes of the same tree, changing it to `Rc`.

(Note: this ignores all push blocking failures!)

Reviewed By: shiqicao

Differential Revision: D16794498

fbshipit-source-id: 830c860e38adab415e24f12975940b4b92aa07ea