Tests for Rust Ppl_class_rewriter
Summary:
Adding comparator option to compare Rust function added in D16794608 to OCaml version.

The idea of this diff stack is that it implements the ugly parts of infrastructure, leaving isolated and well tested function to be implemented... by Not Me. At least not yet - I'll give other people (rest of the team / bootcampers) chance to make some progress on it while working on `Full_fidelity_parse_errors` first.

(Note: this ignores all push blocking failures!)

Reviewed By: shiqicao

Differential Revision: D16793366

fbshipit-source-id: d91d689caa3aad052df905fab2bae1a78699b99a