Vomit 'a all over the codebase even more
Summary:
This is necessary to make SourceText that will be able to become part of the token (see {D16794527}) have any usable lifetime. The diff is created by annotating `LexableToken` with `'a` and fixing all the consequences of it.

Note that with previous diff, {D16794498}, we are one step closet to make the SourceText own the text and get rid of all the lifetimes. This would come at the cost of few copies of entire source text (i.e. when passing it from OCaml, we can't own it - it's owned by OCaml GC). I predict that we will eventually do it, because the simplification will be worth it, and perf hits are unobvious (i.e. we already copy entire source texts between HHVM and hackc workers, and it doesn't seem to hurt). For now, sticking with lifetimes though.

(Note: this ignores all push blocking failures!)

Reviewed By: shiqicao

Differential Revision: D16794512

fbshipit-source-id: 3f9a217734f6192b250d128fdac9882d72cdb588