Builtin function to allow rollout gating in WWW
Summary: I plan on rolling the native autoloader out via a config option in `config.hdf` and `cli.hdf`. See D16832639 for the corresponding opsfiles diff. But the native autoloader stops running on a given request if `autoload_set_paths` is called. I plan on exposing a builtin to WWW which allows it to check whether the native autoloader is running. If so, we can avoid calling `autoload_set_paths`.

Reviewed By: dneiter

Differential Revision: D16838486

fbshipit-source-id: 886f48c456b6b1fe34e86ee828e49478423afc4c