Add Herr for clear naming errors so they end up as Terr
Summary:
`Tany` is inherently unsound, behaving as the sub and super-type of all other types.

This adds a `Herr` to distinguish things that are untyped because of some naming error that may be suppressed later, or because we're in partial mode and things are simply untyped. The end result is that the typechecker sees `Terr` rather than `Tany` and we can better trace sources of unsuppressed unsoundness.

Reviewed By: vassilmladenov

Differential Revision: D16551205

fbshipit-source-id: 375e501a24dd72bfe4fb90e14c78b25cfd1f323b