Remove dead code from cgCall()
Summary:
This optimization is unused and incorrect.

The only case where func->arFuncPtr() is non-null and func->nativeFuncPtr() is
null is when the underlying C++ implementation is undefined, which is not worth
optimizing as it never occurs.

It is also incorrect, as it performs only a small subset of prologue logic,
skipping things such as callee-side dynamic call checks. The only thing it does
is checking surprise flags on enter, which makes situation worse, as there is
no corresponding check for surprise flags on exit.

There's no need to fix it and make it work with builtins, as the Call opcode
with known callee is emitted only when we known the callee, which means we
tried to perform inlining, which would inline NativeImpl using proper
mechanisms.

So, let's just nuke it.

Reviewed By: ricklavoie

Differential Revision: D16909061

fbshipit-source-id: 2a0509e4940909ff66e0270ce396bb38b96fca06