Move call target profiling from Call to ProfileCall IR opcode
Summary:
Move the logic responsible for profiling unknown calls from the Call opcode to
the new ProfileCall opcode.

Slightly improves situation for calls with argument unpacking, which use the
CallUnpack opcode, as JIT will now be able to infer the return type (CallUnpack
doesn't yet take advantage of the known callee).

Makes it possible to remove the prologue guards in the next diff.

Reviewed By: ricklavoie

Differential Revision: D16899505

fbshipit-source-id: c196cbc035383d01ed73c7da617a5a967cf58a43