Reimplement Syntax Tree in Rust
Summary:
Reimplement `full_fidelity_syntax_tree.ml` in Rust. Note that there are some methods that are missing implementation in `syntax.rs` so the `to_json` and `error` methods will need to be added in a following diff when those are implemented.

The implementation of `dedup_errors` and `process_errors` seems a bit excessive given that we can accomplish the same thing by mutating the original vector of errors, but I wanted to maintain those two as pure functions given the original OCaml implementation would be pure functions.

Reviewed By: dabek

Differential Revision: D16797155

fbshipit-source-id: 3160362eefcea8ce33eeaf859bdad2083a9ba25e