Implement fb_intercept2
Summary:
This diff adds `fb_intercept2` which will be a modern replacement to `fb_intercept`.
The original motivation behind this function is `fb_intercept` is used to auto-mock functions but it is unable to handle reified functions, it is not extendible to handle reified functions either.
This replacement invokes a user handler upon calling a function or a class method. This handler is expected to have signature similar to `intercept_handler($name, $obj, $params)` where each argument is same as fb_intercept. This handler is expected to return a shape where if 'value' is given, then value is returned as the result of the original function, if 'callback' is given, then the callback is called as the result of the original function, if 'prepend_this' is given, then this or lsb class is prepended as the first argument to the callback, if neither value or callback is given, then the original function is executed, if the function does not return a shape, then a runtime exception is raised.
Signature of the callback and the original function is required to be the same including the arity and parity of reified arguments.

Reviewed By: jano

Differential Revision: D16871337

fbshipit-source-id: 7cc51c45fa41e9ee89b507e9cb9067949fc416d8