impl Ocamlvalue for SyntaxError
Summary: This, as the change to `TARGETS` file shows, causes `parser_core_types` to depend on things that are OCaml related, which I thought would be a problem for non-OCaml users of this crate (see [this post](https://fb.workplace.com/groups/hhvm.users/permalink/2435105353204732/) for previous instance of such problem). But it seems to work this time? See inline comment on next diff (D16845845) for an example where it does not, and I don't really see what is the difference that triggers it.

Reviewed By: shiqicao

Differential Revision: D16845841

fbshipit-source-id: 2f641d74597b374c665caf391f91e79a060ece84