Make use of pre-IterNext AssertCell
Summary:
In repo mode, we usually generate an AssertType<InitCell> for the iterator output locals right before the IterNext.

We can take advantage of this bytecode generation to remove some checks. We punt on IterNext if we don't see these assertions beforehand; the interpreter handles refs in these positions.

(Note: this ignores all push blocking failures!)

Reviewed By: jjgriego

Differential Revision: D16535712

fbshipit-source-id: a31961855d09b63c5eb9f44fb7959acebeaf1cc1