Fix `provTagHere` to actually reference the srcLoc table
Summary:
Before, we'd shove in the _index_ to this table instead of consulting
it, which is .... not ideal. Trait flattening can mean that some of the
indicies can be -1 since their source unit differs from the function's
originalUnit; I don't think these will be a problem since we'd have seen
logging with line numbers of -1 already and we'll put up a diff to have this
generate a more coherent tag than `file:-1`.

Reviewed By: mxw

Differential Revision: D16954190

fbshipit-source-id: 7f31194826dbfe28ba1fe745dba69a7b7782ab2d