Allow StackLimit refs be shared by threads cheaply
Summary:
This generalizes their usage in a way that one thread
can configure a stack limit and hand it off as "read-only" reference
to the parser without requiring the parser to parse in the same thread
(which is future-proof when parsing becomes parallel).

Consequently, this means it is now possible to catch panics during parsing
and inspect if the stack overflowed from the thread which initiated parsing
(note: StackLimit internally uses lock-free but shared per-thread state).

Reviewed By: dabek

Differential Revision: D16925108

fbshipit-source-id: 164dc8e8381ec20c822e5cb1a3a3f80bceb93dd9