Fix over-assertion in array-iterator
Summary:
I had checkInvariants assert that we used an iteration helper specific to a given array layout.

This assertion is incorrect for local iterators. When we initialize an iterator, we always choose a next helper specific to that array type, but (for local iterators, which don't bump the base ref-count), the base can be COWed and promoted mid-iteration. I've modified checkInvariants to account for this possibility. We still make the stricter assertion for non-local iterators.

I've also removed the assertions in the various IterNext vtable methods. These assertions are all subsumed by checkInvariants, which is called by the methods that dispatch to the vtable methods (e.g. by iter_next_ind and friends). Specifically, the vtable methods used to assert:
- That the iterator type was TypeArray / TypeIterator
- That m_data was set for non-local iterators, and null for local iterators
- That the array had the appropriate layout (accounting for the bug that this diff fixes)

checkInvariants does all this, plus additional assertions on pos and end.

Reviewed By: mxw

Differential Revision: D16968135

fbshipit-source-id: 07c119c228c25e90ebaa7345286b6c4921fda89f