Fix issue with missing canon names.
Summary: So technically, a failure to find a canonical name when we can find a location should always be a bug, since if the file was deleted we should have picked up on that and tombstoned it, but there are occasional race conditions or naming table saved state issues that can cause this to happen.  Instead of hard failing, we should handle this gracefully.

Differential Revision: D16907900

fbshipit-source-id: f124db4c813944e874a9b6b5fedd9e8f3082cfb5