Define string message for func and class conversion to string
Summary: This diff is not supposed to change any behavior. I've just been working on error elevator for functions being converted to string and noticed that one string literal appears in few places in a code and thought it would be better to define it in one place and use it.

Reviewed By: ricklavoie

Differential Revision: D16992811

fbshipit-source-id: a3238a08f3684b808cecfdf9392b28c05c9ff796