Add layer of indirection to construct Tany
Summary:
This (and other diffs in this area) are in preparation for elimination of `Tany` from the type system in favor of `Tdynamic`.

This adds a dummy argument to `Tany` to make it sufficiently difficult to construct them outside of `Typing_defs.make_tany`. By doing this, in the future we will have one place where we can branch on the relevant options to turn all `Tany` into `Tdynamic`. Additionally, all match expressions are updated to account for `Tany`'s argument, so future work in this area will be less painful.

NB: This still doesn't totally prevent people from creating raw `Tany`, but it should be sufficiently difficult enough to indicate that `make_tany` is the recommended way to do it. I think the only way to hard enforce it at the compiler level is to add an MLI for Typing_defs and exclude the TanySentinel module.

Differential Revision: D16923938

fbshipit-source-id: f3839f7d3b4ea235391673df0680ab670da5dc81