Default rust = true in more places
Summary:
D16806756 enabled the local config flag for everyone, but the flag is not threaded everywhere. This can be confusing, for example when your `hh_server` and `hh_single_type_check` binaries default to different parsers. This diff changes the default in more places.

The difference between internal and external default is for the same panic/abort issues as in D17052527.

`Build_id` is only set by `Buck`, so it's an OK approximation of "is this internal FB build" (let me know of a better one available to OCaml codebase). It will be "wrong" in `dune` build, but IMHO this is less important use case than where we are wrong now.

Reviewed By: shiqicao

Differential Revision: D17059871

fbshipit-source-id: dbfcb2b5f49fc5ea5cb7ec0e690f814111d34e11