Sort object properties by countability
Summary:
This diff is a simple change that modifies property sorting to sort properties lexicographically by countedness, then hotness, then logical slot. (Previously, we sorted by hotness, then logical slot.)

Properties that may be counted come first, followed by properties that must be uncounted. Within each group, we try to put hotter properties first. Doing so gives us a branch miss win and a small CPU win because we can set a smaller property index threshold in ObjectData::release.

There are actually three classes of properties: properties which must be heap-allocated (e.g. string), properties which may be heap allocated (e.g. string | null), and properties which are uncounted. If we distinguished between all three of these classes, we wouldn't even have to check the type byte for the first class in release. I leave that as a follow-up. It's a bit trickier than this change because objects start out in an uninitialized state, which we have to check for.

Reviewed By: ottoni

Differential Revision: D16844781

fbshipit-source-id: ef66c0bbcec035742ab10906fbc7b4b6dc8bff17