Lock down the Iter interface
Summary:
ArrayIter is the iterator class used by HHVM C++ code to iterate over "stuff": arrays, collections, Objects implementing Iterator. It supports all sorts of wild features like copying, moving, resetting, getting rvals (and thus getting lvals), etc.

Iter is the iterator class exposed by HHVM to Hack through the interpreter and JIT. It supports forward iteration in two slightly different modes: in "foreach" loops (with the IterInit / IterNext family of bytecodes) and in delegated generators. Iter happens to share a memory layout with ArrayIter, because they are similar.

Unfortunately, there's a big ol' hole in the Iter API: a public method that returns the underlying ArrayIter. We replace this API with a more constrained two-part API:
- A complete set of methods on the Iter class for generic iteration
- A handful of additional helpers that basically implement IterInit / IterNext w/ a single vtable lookup.

We modify the interpreter to use this new API; the JIT was always using it. As a nice side effect of these changes, we no longer need the ugly "escalate" method, since the interpreter now uses the same logic for IterNext as the JIT. Doing so is a significant performance win. I'm not sure if that's because we're no longer despecializing ArrayMixedNoTombstones iterators, or if just optimizing the interpreter iterator call has a big effect.

There's one more ugly thing about this API: the fact that key() and val() return Variant. I want to fix that in a follow-up, because usage of Variant in this file is widespread and unnecessary. Once I figure out the correct sequence of actual ops to replace Variant with, I'll do so throughout array-iterator.

Reviewed By: mxw

Differential Revision: D16924031

fbshipit-source-id: 8b5cdcc4a6e80fdb689d0129610057c7d6ecc5f8