Fix unlikely memory leak in arrprov::tagTV()
Summary:
If a builtin returns a multiply-referenced vec or dict that does not
have a provenance tag, tagTV() copies it but does not, in its current
implementation, decref it to compensate for the reference that's being
dropped in favor of tagTV()'s return value.

Fix this, and additionally refactor tagTV() slightly for immediate use
in stacked diffs.

Reviewed By: jjgriego

Differential Revision: D17085876

fbshipit-source-id: 92d671cf503fd3c6f955c8d675d1235a4b1ef7e0