Improve rematerialization in vasm-graph-color
Summary:
Improve the rematerialization logic in vasm-graph-color. Instead of
trying to calculate all the possible rematerialization candidates in
one pass, calculate it on demand. That is, at a given program point
where a reload is needed, we walk backwards looking for the def of the
Vreg being reloaded. Once we find it, we examine the instruction and
check if its suitable for rematerialization (its pure and all of its
sources are available). If found, we use that instead of a reload. To
avoid overhead, we cache the found defining instruction.

This is more efficient because we only run the logic when we're
actually doing a reload (rather than for every single Vreg). Its also
more flexible, and lets us incorporate better rematerialization
criteria. For example, with this, we can now permit rematerializing
instructions which have sources.

Reviewed By: ottoni

Differential Revision: D17084430

fbshipit-source-id: 0b6c81e512905a6b61d333883ee79964dd49490f