Rework spilling heuristic in vasm-graph-color
Summary:
Rework the spilling heuristic in vasm-graph-color. The old heuristic
is a pretty traditional one: choose the Vreg which is used farthest
away from the current location. This works, but we can do better using
the profiling information already available. Instead of using the
distance, use the profiling information to estimate how many reloads
we'll have to execute at runtime.

More specifically, for a specific Vreg at a specific location, look
forward in the unit for all the immediate uses of that Vreg. We know
that if we spill that Vreg, we'll have to emit a reload before all of
those uses. Use the profiled execution counts on the blocks where the
uses are to determine how many times those reloads will be executed at
runtime. Once the Vreg is reloaded, it won't be spilled anymore, so we
don't advance past any use of the Vreg. Since profiling information
can be coarse at times, we also calculate the distance and use it as a
tie breaker. The Vregs with the lowest usage and higher distance are
choosen to be spilled.

Copyish instructions cause some complications. They generally do not
trigger a reload (they can copy spilled Vregs). So, if we encounter
those, we do not consider it a use. However the dest of the copy is
now also a spill, so we now also look for uses of that Vreg. So, in
reality, we don't look for uses of a single Vreg, but a set. Once a
Vreg is used, we drop it from the set. We're done once the set is
empty.

This approach has some advantages. Not only does it generate better
code, but its generally more efficient. We now calculate the spill
weights on demand instead of pre-calculating it for all Vregs. Since a
lot of Vregs will never be considered for spilling, this saves work.

We also use the profiled execution counts in one other place. When
calculating the initial spiller state for a block, we'll sum up the
block weights of all the predecessors, broken down into non-spilled
and spilled for each Vreg. If the Vreg is in a majority of spilled,
we'll treat it as initially spilled, and vice-versa.

Differential Revision: D17067630

fbshipit-source-id: c9528f52e37be9ced7e8d3116ff9fac0e95248fa